Add turbo mode, extend difficulty, and fixes
Introduce a single-player turbo mode (g_turboMode) that scales player, vehicle and mover speeds and adjusts mover time granularity; disable weapon reload behavior while turbo is active. Extend skill system to 5 levels (MAX_SKILL_LEVELS=5), adjust difficulty/damage scaling and related AI aggression calculations. Improve portability and correctness for pointer-sized data: replace dword*/int casts with uintptr_t/intptr_t in Heap and other code, fix alignment/Allocate16 logic, and tighten asserts in AAS_routing. Enhance event save/restore to support typed argument serialization (including 64-bit integer args and trace material names). Misc: rename flashlight console command to actorFlashlight, add debug warnings/prints for client moveable spawn failures, change importDLL to intptr_t, fix monster death effect call, and update README/meson_options to reflect x64 build behavior and artifact naming.

+++ src/idlib/containers/Queue.h
@@ -2,6 +2,8 @@
 #ifndef __QUEUE_H__
 #define __QUEUE_H__
 
+#include <stddef.h>
+
 /*
 ===============================================================================
 
@@ -64,7 +66,7 @@ type *idQueue<type, nodePtr>::RemoveFirst( void ) {
 	return element;
 }
 
-template< class type, int nextOffset >
+template< class type, size_t nextOffset >
 class idQueueTemplate {
 public:
 							idQueueTemplate( void );
@@ -79,12 +81,12 @@ private:
 
 #define QUEUE_NEXT_PTR( element )		(*((type**)(((byte*)element)+nextOffset)))
 
-template< class type, int nextOffset >
+template< class type, size_t nextOffset >
 idQueueTemplate<type,nextOffset>::idQueueTemplate( void ) {
 	first = last = NULL;
 }
 
-template< class type, int nextOffset >
+template< class type, size_t nextOffset >
 void idQueueTemplate<type,nextOffset>::Add( type *element ) {
 	QUEUE_NEXT_PTR(element) = NULL;
 	if ( last ) {
@@ -95,7 +97,7 @@ void idQueueTemplate<type,nextOffset>::Add( type *element ) {
 	last = element;
 }
 
-template< class type, int nextOffset >
+template< class type, size_t nextOffset >
 type *idQueueTemplate<type,nextOffset>::Get( void ) {
 	type *element;
