+++ src/mpgame/anim/Anim_Import.cpp
@@ -16,7 +16,7 @@ static idStr				Maya_Error;
 
 static exporterInterface_t	Maya_ConvertModel = NULL;
 static exporterShutdown_t	Maya_Shutdown = NULL;
-static int					importDLL = 0;
+static intptr_t				importDLL = 0;
 
 bool idModelExport::initialized = false;
